LAB5/requirements.txt
#wspólne biblioteki
numpy>=1.26,<2.0
pillow>=10,<13
matplotlib>=3.8,<3.11
scikit-learn>=1.3,<1.8

#tensorFlow – CPU (Win) / GPU (Linux)
tensorflow==2.17.0; platform_system == "Windows"
tensorflow[and-cuda]==2.17.0; platform_system == "Linux"
